+++ app/src/main/java/elite/intel/ui/view/PlayerTabPanel.java
@@ -127,7 +127,37 @@ public class PlayerTabPanel extends JPanel {
         btns.add(conversationModeCheckBox);
 
         useVmCheckBox = new JCheckBox("Delegate Controller to VM");
-        useVmCheckBox.addActionListener(e -> playerSession.setUseVm(useVmCheckBox.isSelected()));
+        useVmCheckBox.addActionListener(e -> {
+            if (useVmCheckBox.isSelected()) {
+                JOptionPane.showMessageDialog(
+                        this,
+                        """
+                                <html>
+                                <h3>Experimental Integration</h3>
+                                <p>
+                                    This will disable Elite Intel internal keyboard controller
+                                    and delegate to VoiceMacro, a Windows Only application.
+                                </p>
+                                <p>
+                                    The functionality will be limited to basic commands only.
+                                </p>
+                                <ul>
+                                <li>Advanced options such as route plotting will not be possible.</li>
+                                <li>All handicap features will be disabled.</li>
+                                <li>The app WILL NOT operate keyboard controls at all.</li>
+                                </ul>
+                                <p>All LLM command actions will be sent to VoiceMacro. </p>
+                                <p>You must setup 'EliteDangerous' profile and name keyboard macros matching the action names exactly</p>
+                                <p>See Commands.java class on the GitHub for current list of commands</p>
+                                <p>This list will change without notice</p>
+                                </html>
+                                """,
+                        "Experimental Feature",
+                        JOptionPane.WARNING_MESSAGE
+                );
+            }
+            playerSession.setUseVm(useVmCheckBox.isSelected());
+        });
         btns.add(useVmCheckBox);
 
         add(btns, gbc);